+++ app/models/collected-metadatum.ts
@@ -9,13 +9,7 @@ import tuple from 'ember-osf-web/utils/tuple';
 import Collection, { ChoicesFields } from './collection';
 import Node from './node';
 import OsfModel from './osf-model';
-import { SubjectRef } from './taxonomy';
 import User from './user';
-
-export interface DisplaySubject {
-    text: string;
-    path: string;
-}
 
 export const choiceFields = tuple(
     'collectedType',
@@ -45,17 +39,6 @@ const Validations = buildValidations({
             ],
         };
     }, {}),
-    subjects: [
-        validator('presence', {
-            presence: true,
-            minLength: 1,
-            messageKey: 'validationErrors.min_subjects',
-        }),
-        validator('array', {
-            minLength: 1,
-            messageKey: 'validationErrors.min_subjects',
-        }),
-    ],
 });
 
 export default class CollectedMetadatumModel extends OsfModel.extend(Validations) {
@@ -63,31 +46,11 @@ export default class CollectedMetadatumModel extends OsfModel.extend(Validations
     @attr('string') issue?: string;
     @attr('string') programArea?: string;
     @attr('string') status?: string;
-    @attr('subjects') subjects!: SubjectRef[][];
     @attr('string') volume?: string;
 
     @belongsTo('collection') collection!: Collection;
     @belongsTo('node') guid!: Node;
     @belongsTo('user') creator!: User;
-
-    @computed('subjects')
-    get displaySubjects(): DisplaySubject[] {
-        // returns a list of unique subjects and its path for display and filtering
-        const displaySubjects: DisplaySubject[] = [];
-        this.subjects.forEach(subjectPathArray => {
-            let index = 0;
-            const includedTester = (element: DisplaySubject) => element.text === subjectPathArray[index].text;
-            for (index = 0; index < subjectPathArray.length; index++) {
-                // Only append the subjects if they are not already included in the list
-                if (!displaySubjects.some(includedTester)) {
-                    const { text } = subjectPathArray[index];
-                    const path = ['', ...subjectPathArray.slice(0, index + 1).map(item => item.text)].join('|');
-                    displaySubjects.push({ text, path });
-                }
-            }
-        });
-        return displaySubjects;
-    }
 
     @computed('collection.displayChoicesFields.[]')
     get displayChoiceFields() {