+++ CHANGELOG.md
@@ -10,10 +10,35 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `form-controls` - a form-input wrapper that takes a changeset
 
 ### Changed
+- Models
+    - `collected-metadatum`
+        - removed `subjects` validation
+        - removed `subjects` attribute
+        - removed `displaySubjects` computed
 - Components
     - `project-contributors/list`
         - add ability to load more pages of contributors
         - add loading indicator
+- Engines
+    - `collections`
+        - Components
+            - `collections-submission` - removed subjects section
+            - `discover-page` - removed subject/taxonomy search filter
+            - `collection-search-result` - removed subjects display
+        - Routes
+            - `discover` - removed subject/taxonomy facet
+
+### Removed
+- Components
+    - `subject-picker`
+- Engines
+    - `collections`
+        - Components
+            - `collection-subjects-list`
+            - `discover-page/facets/taxonomy`
+            - `discover-page/active-filters/taxonomy`
+        - Helpers
+            - `custom-taxonomy-filter`
 
 ## [19.7.1] - 2019-08-05
 ### Changed